+++ src/integrations/differentialequations.jl
@@ -42,7 +42,6 @@ function getiterator(sol::DiffEqBase.DESolution)
     end
     
     t_expr = NamedTuples.make_tuple(col_expressions)
-    t_expr.args[1] = Expr(:., :NamedTuples, QuoteNode(t_expr.args[1]))
 
     t2 = :(DESolutionIterator{Float64,Float64})
     t2.args[2] = t_expr