+++ NEWS.md
@@ -1,2 +1,17 @@
+# IterableTables.jl v0.3.0 Release Notes
+* Use NamedTuples.jl v4.0.0
+
+# IterableTables.jl v0.2.0 Release Notes
+* Add Temporal.jl integration
+
+# IterableTables.jl v0.1.0 Release Notes
+* Drop julia 0.5 compat
+* Fix julia 0.6 compat issues
+* Use DataValue for missing values
+
+# IterableTables.jl v0.0.2 Release Notes
+* Add documentation
+* Small bug fixes
+
 # IterableTables.jl v0.0.1 Release Notes
 * Initial release